+++ src/CloudBase.php
@@ -10,4 +10,9 @@ class CloudBase
     {
         return sprintf('%s/views', $_ENV['APP_BASE_PATH']);
     }
+
+    public function getCoreLayout(string $layout = null): string
+    {
+        return sprintf('%s/views/layout/%s.layout.latte', $_ENV['APP_BASE_PATH'], $layout ?? 'cloudbase');
+    }
 }