+++ src/Controller/CloudBaseController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace CloudBase\PluginCore\Controller;
 
 use CloudBase\PluginCore\Classes\Latte\EngineBuilder;
+use CloudBase\PluginCore\CloudBase;
 use Latte\Engine;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -12,9 +13,9 @@ class CloudBaseController
 {
     private Engine $latteEngine;
 
-    public function __construct(string $packageName = null)
+    public function __construct()
     {
-        $this->latteEngine = (new EngineBuilder())->getEngine($packageName);
+        $this->latteEngine = (new EngineBuilder())->getEngine();
     }
 
     protected function renderedLatteResponse(string $template, array $data = []): Response
@@ -23,6 +24,8 @@ class CloudBaseController
             $template .= '.latte';
         }
 
-        return new Response($this->latteEngine->renderToString($template, $data));
+        return new Response($this->latteEngine->renderToString($template, array_merge($data, [
+            'app' => new CloudBase(),
+        ])));
     }
 }