+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CloudBase | Plugin Core | Changelog
 
+## [1.2.0] - 2025-10-06
+### Added
+- Pass `CloudBase` instance to all Latte templates via the `$app` variable.
+- Allow plugins to hook into CloudBase core layouts with `$app->getCoreLayout($layout)`.
+
 ## [1.1.2] - 2025-10-06
 ### Added
 - Add license to `composer.json`.